essentials v1.0.1
Home page: http://essentials.seadex.de
Documentation: http://essentials.seadex.de/Documentation.html
License: http://essentials.seadex.de/License.html
