+++ essentials/include/essentials/base.hpp
