+++ essentials/include/essentials/exception.hpp
