+++ essentials/include/essentials/exception_template.hpp
