+++ essentials/include/essentials/non_copyable.hpp
